feat: 요청 첫번째 라인 받기 테스트 수정 완료
